Factor out ExperienceByTypeListing w defaultOpen
